refactor: rename `policies` to `other_policies`
