+++ python/scylla/_rust/session_builder.pyi
@@ -3,8 +3,8 @@ from __future__ import annotations
 from collections.abc import Sequence
 from ipaddress import IPv4Address, IPv6Address
 
-from .policies import AuthenticatorProvider, AddressTranslator, TimestampGenerator, HostFilter
 from .execution_profile import ExecutionProfile
+from .other_policies import AddressTranslator, AuthenticatorProvider, HostFilter, TimestampGenerator
 from .session import Session
 
 ContactPoint = str | tuple[str | IPv4Address | IPv6Address, int]
