+++ src/lib.rs
@@ -17,7 +17,7 @@ mod deserialize;
 mod enums;
 mod errors;
 mod execution_profile;
-mod policies;
+mod other_policies;
 mod routing;
 mod serialize;
 mod session;
@@ -66,7 +66,7 @@ fn scylla(py: Python<'_>, module: &Bound<'_, PyModule>) -> PyResult<()> {
     add_submodule(py, module, "types", types::types)?;
     add_submodule(py, module, "value", value::value)?;
     add_submodule(py, module, "batch", batch::batch)?;
-    add_submodule(py, module, "policies", policies::policies)?;
+    add_submodule(py, module, "other_policies", other_policies::other_policies)?;
     add_submodule(py, module, "cluster", cluster::cluster)?;
     add_submodule(py, module, "routing", routing::routing)?;
     Ok(())
